% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate data by categorical variables}
\usage{
aggregate_data(
  data,
  group_vars,
  summaries,
  vars = NULL,
  names = NULL,
  quantiles = c(0.25, 0.75)
)
}
\arguments{
\item{data}{a dataframe or survey design object to aggregate}

\item{group_vars}{a character vector of categorical variables in \code{data}
to group by}

\item{summaries}{summaries to generate for the groups generated
in \code{group_vars} (see details)}

\item{vars}{names of variables in the dataset to calculate summaries of}

\item{names}{name templates for created variables (see details)}

\item{quantiles}{if requesting quantiles, specify the desired quantiles here}
}
\value{
aggregated dataframe containing the summaries
with tidyverse code attached
}
\description{
Aggregate a dataframe into summaries of all numeric/date-time variables by
grouping them by specified categorical variables
and returns the result along with tidyverse code used to generate it.
}
\section{Calculating variable summaries}{

The \code{aggregate_data} function accepts any R function which returns a
single-value (such as \code{mean}, \code{var}, \code{sd}, \code{sum}, \code{IQR}). The default name of
new variables will be \verb{\{var\}_\{fun\}}, where \code{{var}} is the variable name and
\code{{fun}} is the summary function used. You may pass new names via the \code{names}
argument, which should be either a vector the same length as \code{vars}, or a
named list (where the names are the summary function). In either case, use
\code{{var}} to represent the variable name. e.g., \verb{\{var\}_mean} or \verb{min_\{var\}}.

You can also include the summary \code{missing}, which will count the number of
missing values in the variable. It has default name \verb{\{var\}_missing}.

For the \code{quantile} summary, there is the additional argument \code{quantiles}.
A new variable will be created for each specified quantile 'p'. To name these
variables, use \code{{p}} in \code{names} (the default is \verb{\{var\}_q\{p\}}).
}

\examples{
aggregated <-
    aggregate_data(iris,
        group_vars = c("Species"),
        summaries = c("mean", "sd", "iqr")
    )
cat(code(aggregated))
head(aggregated)

}
\seealso{
\code{\link{code}}
}
\author{
Tom Elliott, Owen Jin, Stephen Su
}
