% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_dt.R
\name{aggregate_dt}
\alias{aggregate_dt}
\title{A wrapper function of \code{iNZightTools::aggregate_data()} for date-time grouping}
\usage{
aggregate_dt(
  data,
  dt,
  dt_comp,
  summaries,
  vars = NULL,
  names = NULL,
  quantiles = c(0.25, 0.75)
)
}
\arguments{
\item{data}{a dataframe or survey design object to aggregate}

\item{dt}{a string of the name of the date-time variable}

\item{dt_comp}{a string of the component of the date-time to group by}

\item{summaries}{an unnamed character vector or named list (with the names
being the names of variables in the dataset to calculate summaries of,
and the elements being character vectors) of summaries to generate
for the groups generated in \code{group_vars}}

\item{vars}{names of variables in the dataset to calculate summaries of
(ignored if \code{summaries} is a named list)}

\item{names}{name templates for created variables}

\item{quantiles}{if requesting quantiles, specify the desired quantiles here}
}
\value{
aggregated dataframe containing the summaries
with tidyverse code attached
}
\description{
A wrapper function of \code{iNZightTools::aggregate_data()} for date-time grouping
}
\seealso{
\code{\link{aggregate_data}}
}
\author{
Zhaoming Su
}
