% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_dt.R
\name{aggregatedt}
\alias{aggregatedt}
\title{Aggregate datetimes}
\usage{
aggregatedt(.data, method, key, name)
}
\arguments{
\item{.data}{dataframe or tibble to aggregate}

\item{method}{the type of aggregation}

\item{key}{the key column}

\item{name}{the name of the variable}
}
\value{
a data frame/tibble
}
\description{
Aggregate datetimes
}
\author{
Yiwen He
}
