% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_rows.R
\name{append_rows}
\alias{append_rows}
\title{Append rows to the dataset}
\usage{
append_rows(data, new_data, when_added = FALSE)
}
\arguments{
\item{data}{original dataset}

\item{new_data}{dataset to append below \code{data}}

\item{when_added}{whether a \code{.when_added} column is required}
}
\value{
dataset with new rows appended
}
\description{
Append rows to the dataset
}
\author{
Yiwen He, Zhaoming Su
}
