% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_rows.R
\name{appendrows}
\alias{appendrows}
\title{Append row to the dataset}
\usage{
appendrows(.data, imported_data, date)
}
\arguments{
\item{.data}{original dataset}

\item{imported_data}{imported dataset}

\item{date}{whether a "When_Added“ column is required}
}
\value{
dataset with new rows appended
}
\description{
Append row to the dataset
}
\author{
Yiwen He
}
