% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{changeColumnTypes}
\alias{changeColumnTypes}
\title{Changes the datatype of columns of dataframe to the specified datatypes.}
\usage{
changeColumnTypes(user.data.frame, col.types)
}
\arguments{
\item{user.data.frame}{dataframe which needs the datatypes of columns changed.}

\item{col.types}{a vector of datatypes of columns.}
}
\value{
a data frame.
}
\description{
\code{changeColumnTypes} returns a data frame with changed datatypes of columns. 
Returns the same data frame if column types are specified as NULL.
}

