% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{changeToDatatype}
\alias{changeToDatatype}
\title{Changes the datatype of a column of a dataframe to the specified datatype.}
\usage{
changeToDatatype(user.data.frame, x = 0, col.types = NULL)
}
\arguments{
\item{user.data.frame}{dataframe which needs the datatypes of columns changed.}

\item{x}{column number of the column which needs the data type changed.}

\item{col.types}{a vector of datatypes of columns.}

\item{...}{Other arguments.}
}
\value{
a data frame.
}
\description{
\code{changeToDatatype} returns a data frame with 1 column
(just the column whose data type has been changed) with changed
datatypes of column.
}

