% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{checkForColumnNames}
\alias{checkForColumnNames}
\title{Checks if all the column names specified in the metadata 
are present in the header line or not.}
\usage{
checkForColumnNames(metadata.list, first.line)
}
\arguments{
\item{metadata.list}{Metadata passed as a list.}

\item{first.line}{A string. First line which is not part of the metadata.}
}
\value{
A logical vector of length 1.
}
\description{
\code{checkForColumnNames} Returns TRUE if all the columns specified 
in the metadata are present in the first line which is not part of 
the metadata.
}

