% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_writing.R
\name{code}
\alias{code}
\title{Get Data's Code}
\usage{
code(data)
}
\arguments{
\item{data}{dataset you want to extract the code from}
}
\value{
The code used to generate the data.frame, if available (else NULL)
}
\description{
Used to grab code from a data.frame generated by this package.
}
\details{
This is simply a helper function to grab the contents
of the `code` attribute contained in the data object.
}
\author{
Tom Elliott
}
