% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{collapse_cat}
\alias{collapse_cat}
\title{Collapse data by values of a categorical variable}
\usage{
collapse_cat(data, var, collapsed = NULL, name = NULL)
}
\arguments{
\item{data}{a dataframe to collapse}

\item{var}{a string of the name of the categorical variable to collapse}

\item{collapsed}{a named list: the values of the categorical variable
matching each element in the list will be changed to the name of that
element in the list}

\item{name}{a name for the new variable}
}
\value{
the original dataframe containing a new column of the
        collapsed variable with tidyverse code attached
}
\description{
Collapse values in a categorical variable into less levels of values
}
\examples{
collapsed <- collapse_cat(iris,
    var = "Species",
    list(S = "setosa", V = c("versicolor", "virginica"))
)
cat(code(collapsed))
head(collapsed)

}
\seealso{
\code{\link{code}}
}
\author{
Stephen Su
}
