% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.levels.R
\name{combine.levels}
\alias{combine.levels}
\title{Combine the levels of factor variables.}
\usage{
combine.levels(dafr, columns)
}
\arguments{
\item{dafr}{The data.frame containing factor columns 
specified in columns.}

\item{columns}{The column names of the columns in dafr 
to combine.}
}
\value{
A data.frame with one additional column as dafr,
which contains the combined levels.
}
\description{
Combine the levels of factor variables.
}
\author{
Christoph Knapp
}

