% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{combine_cat}
\alias{combine_cat}
\title{Combine categorical variables into one}
\usage{
combine_cat(
  data,
  vars,
  sep = ":",
  name = NULL,
  keep_empty = FALSE,
  keep_na = TRUE
)
}
\arguments{
\item{data}{a dataframe with the columns to be combined}

\item{vars}{a character vector of the categorical variables to be combined}

\item{sep}{a character string to separate the levels}

\item{name}{a name for the new variable}

\item{keep_empty}{logical, if \code{FALSE} empty level combinations
are removed from the factor}

\item{keep_na}{logical, if \code{TRUE} the \code{<NA>} in the factors or
\code{NA} in the characters will be replaced with \code{"<NA>"};
otherwise, the resulting entries will return \code{<NA>}}
}
\value{
original dataframe containing new columns of the renamed
        categorical variable with tidyverse code attached
}
\description{
Combine specified categorical variables by concatenating
their values into one character, and returns the result
along with tidyverse code used to generate it.
}
\examples{
combined <- combine_cat(warpbreaks, vars = c("wool", "tension"), sep = "_")
cat(code(combined))
head(combined)

}
\author{
Owen Jin, Stephen Su
}
