% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cat_var.R
\name{combineCatVars}
\alias{combineCatVars}
\title{Combine categorical variables into one}
\usage{
combineCatVars(.data, vars, sep = ".")
}
\arguments{
\item{.data}{a dataframe with the columns to be combined}

\item{sep}{the seperator to combine the values of the variables in \code{var} by.
"." by default}

\item{var}{a character vector of the categorical variables to be combined}
}
\value{
original dataframe containing a new column of the renamed categorical variable with tidyverse code attached
}
\description{
Combine specified categorical varaibles by concatenating 
their values into one character, and returns the result
along with tidyverse code used to generate it.
}
\examples{
combined <- combineCatVars(warpbreaks, vars = c("wool", "tension"), sep = "_")
code(combined)
head(combined)

}
\seealso{
\code{\link{combine_levels}} ######
}
\author{
Owen Jin
}
