% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{convert_to_cat}
\alias{convert_to_cat}
\title{Convert variables to categorical variables}
\usage{
convert_to_cat(data, vars, names = NULL)
}
\arguments{
\item{data}{a dataframe with the categorical column to convert}

\item{vars}{a character vector of column names to convert}

\item{names}{a character vector of names for the created variables}
}
\value{
original dataframe containing new columns of the
        converted variables with tidyverse code attached
}
\description{
Convert specified variables into factors
}
\examples{
converted <- convert_to_cat(iris, vars = c("Petal.Width"))
cat(code(converted))
head(converted)

}
\seealso{
\code{\link{code}}
}
\author{
Zhaoming Su
}
