% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{convert_to_date}
\alias{convert_to_date}
\title{Convert variables to dates}
\usage{
convert_to_date(data, vars, ord = NULL, names = NULL)
}
\arguments{
\item{data}{a dataframe with the variables to convert}

\item{vars}{a character vector of column names to convert}

\item{ord}{a character vector of date-time formats}

\item{names}{a character vector of names for the created variables}
}
\value{
original dataframe containing new columns of the
        converted variables with tidyverse code attached
}
\description{
Convert variables to dates
}
\seealso{
\code{\link{code}}
}
\author{
Zhaoming Su
}
