% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dt.R
\name{convert_to_datetime}
\alias{convert_to_datetime}
\title{Convert to datetime}
\usage{
convert_to_datetime(data, factorname, convname, newname)
}
\arguments{
\item{data}{dataframe}

\item{convname}{format}

\item{varname}{name of the variable}

\item{name}{name of the new column}
}
\value{
dataframe with datatime column
}
\description{
Convert to datetime
}
\author{
Yiwen He
}
