% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{convertSpssDetails}
\alias{convertSpssDetails}
\title{Converts the data type of columns specified from numeric to date.}
\usage{
convertSpssDetails(user.data.frame, date.columns)
}
\arguments{
\item{user.data.frame}{the dataframe returned after
reading an spss file.}

\item{date.columns}{A numeric vector. Has the integer
values of the columns that have to be converted.}
}
\value{
A data frame
}
\description{
\code{convertSpssDetails} converts the specified columns in
the dataframe returned after reading an SPSS file to datatype
Date from numeric.
}

