% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{convertToLevelsAndLabels}
\alias{convertToLevelsAndLabels}
\title{Extracts the levels and labels to be used when converting a column to factor.}
\usage{
convertToLevelsAndLabels(obj)
}
\arguments{
\item{obj}{}
}
\value{
A list.
}
\description{
\code{convertToLevelsAndLabels} Returns a list with levels vector and labels vector.
}

