% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{convertToMetadata}
\alias{convertToMetadata}
\alias{convertToMetadata.data.frame}
\alias{convertToMetadata.list}
\title{Takes an object (list or a dataframe) and returns its metadata.}
\usage{
convertToMetadata(obj)

\method{convertToMetadata}{data.frame}(user.data.frame)

\method{convertToMetadata}{list}(metadata.list)
}
\arguments{
\item{obj}{The object whose metadata has to be returned.}

\item{user.data.frame}{the dataframe to be converted}

\item{metadata.list}{list of metadata information}
}
\value{
A string - metadata.
}
\description{
\code{convertToMetadata} Returns the metadata for the object passed.
}

