% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_writing.R
\name{create_varname}
\alias{create_varname}
\title{Create variable name}
\usage{
create_varname(x)
}
\arguments{
\item{x}{a string to convert}
}
\value{
a string, which is also a valid variable name
}
\description{
Convert a given string to a valid R variable name,
converting spaces to underscores (_) instead of dots.
}
\examples{
create_varname("a new variable")
create_varname("8d4-2q5")
}
\author{
Tom Elliott
}
