% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{delete_vars}
\alias{delete_vars}
\title{Delete variables}
\usage{
delete_vars(data, vars = NULL)
}
\arguments{
\item{data}{dataset}

\item{vars}{variable names to delete}
}
\value{
dataset without chosen variables
}
\description{
Delete variables from a dataset
}
\author{
Zhaoming Su
}
