% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_vars.R
\name{deleteVars}
\alias{deleteVars}
\title{Delete variables}
\usage{
deleteVars(.data, vars)
}
\arguments{
\item{.data}{dataset}

\item{vars}{variables to delete}
}
\value{
dataset without chosen variables
}
\description{
Delete variables from a dataset
}
\author{
Tom Elliott
}
