% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dt_comp.R
\name{extract_dt_comp}
\alias{extract_dt_comp}
\title{Extract date component from a date-time variable}
\usage{
extract_dt_comp(data, var, comp, name = NULL)
}
\arguments{
\item{data}{dataframe}

\item{var}{name of the date-time variable}

\item{comp}{date component wanted from the variable, see
\code{iNZightTools:::inz_dt_comp} for full list}

\item{name}{name of the new column}
}
\value{
dataframe with the new date component column
}
\description{
Extract date component from a date-time variable
}
\author{
Zhaoming Su
}
