% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dt_comp.R
\name{extract_part}
\alias{extract_part}
\title{Extract part of a datetimes variable (DEPRECATED)}
\usage{
extract_part(.data, varname, part, name)
}
\arguments{
\item{.data}{dataframe}

\item{varname}{name of the variable}

\item{part}{part of the variable wanted}

\item{name}{name of the new column}
}
\value{
see `extract_dt_comp`
}
\description{
This function has been replaced by `extract_dt_comp` and will be removed in the next release.
}
