% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{extractFromFrame}
\alias{extractFromFrame}
\alias{extractFromFrame.iNZclass.factor}
\alias{extractFromFrame.iNZclass.numeric}
\alias{extractFromFrame.iNZclass.character}
\title{Extracts more information from the dataframe (obj).}
\usage{
extractFromFrame(obj, metadata.list)

\method{extractFromFrame}{iNZclass.factor}(obj, metadata.list)

\method{extractFromFrame}{iNZclass.numeric}(obj, metadata.list)

\method{extractFromFrame}{iNZclass.character}(obj, metadata.list)
}
\arguments{
\item{obj}{A vector.}

\item{metadata.list}{A list with information about the metadata of the object.}
}
\value{
A list.
}
\description{
\code{extractFromFrame} Returns a list.
}
