% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R
\name{filterRows}
\alias{filterRows}
\title{Filter data by row numbers}
\usage{
filterRows(.data, rows)
}
\arguments{
\item{.data}{a dataframe or a survey design object to filter}

\item{rows}{a numeric vector of row numbers to slice off}
}
\value{
filtered dataframe with tidyverse code attached
}
\description{
Filter a dataframe by slicing off specified rows
and returns the result along with tidyverse code used to generate it.
}
\examples{
filtered <- filterRows(iris, rows = c(1,4,5))
cat(code(filtered))
head(filtered)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin
}
