% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{findColumnNames}
\alias{findColumnNames}
\title{Reads a file and returns the first line which isn't a part of the metadata.}
\usage{
findColumnNames(path)
}
\arguments{
\item{path}{A string. Specifies the location of the file to be read.}
}
\value{
A string.
}
\description{
\code{findColumnNames} Returns the first line which isn't a part of metadata.
}

