% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{fitDesign}
\alias{fitDesign}
\title{Fit a survey design}
\usage{
fitDesign(svydes, dataset.name)
}
\arguments{
\item{svydes}{a design}

\item{dataset.name}{a dataset name}
}
\value{
a survey object
}
\description{
Fit a survey design to an object
}
\author{
Tom Elliott
}
