% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfitting.R
\name{fitModel}
\alias{fitModel}
\title{Fit a Model}
\usage{
fitModel(y, x, data, family = "gaussian", design = "simple", svydes = NA,
  ...)
}
\arguments{
\item{y}{character string representing the response,}

\item{x}{character string of the explanatory variables,}

\item{data}{name of the object containing the data.}

\item{family}{gaussian, binomial, poisson (so far, no others will be added)}

\item{design}{data design specification. one of 'simple', 'survey' or 'experiment'}

\item{svydes}{a vector of arguments to be passed to the svydesign function, excluding data (defined above)}

\item{...}{further arguments to be passed to lm, glm, svyglm, such as offset, etc.}
}
\value{
A fit object (lm, glm, or svyglm)
}
\description{
Fit models
}
\details{
wrapper function
}
\author{
Tom Elliott
}

