% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_levels.R
\name{get.collapsed.column}
\alias{get.collapsed.column}
\title{Collapse data by values of a categorical variable}
\usage{
get.collapsed.column(.data, var, levels)
}
\arguments{
\item{.data}{a dataframe to collapse}

\item{var}{a character of the name of the catgorical variable to collapse}

\item{levels}{a character vector of the levels to be collapsed}
}
\value{
the original dataframe containing a new column of the collapsed variable with tidyverse code attached
}
\description{
Collapse several values in a categorical variable into one level
}
\examples{
collapsed <- get.collapsed.column(iris, vars = "Species", levels = ("versicolor", "virginica")
code(collapsed)
head(collapsed)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin
}
