% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_categorical.R
\name{get.combinations}
\alias{get.combinations}
\title{Takes a data.frame as generated from the 
\code{get.missing.categorical} function and converts it 
into a data.frame of all unique rows and there counts in 
the original data.}
\usage{
get.combinations(dafr)
}
\arguments{
\item{dafr}{A data.frmae to convert}
}
\value{
A data.frame with all unique rows from dafr and 
their counts in the last column.
}
\description{
Takes a data.frame as generated from the 
\code{get.missing.categorical} function and converts it 
into a data.frame of all unique rows and there counts in 
the original data.
}
\author{
Christoph Knapp
}

