% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.form.class.interval.R
\name{get.form.class.interval}
\alias{get.form.class.interval}
\title{Form class intervals from a column specified by column 
name or column index from the a data.frame.}
\usage{
get.form.class.interval(dafr, column, num.intervals,
  open.left.closed.right = T, method = c("equal.width", "equal.count",
  "specified"), intervals = NULL, labels = NULL)
}
\arguments{
\item{dafr}{The data frame the column can be retrieved from.}

\item{column}{The column name or index to produce class 
intervals from.}

\item{num.intervals}{The number of intervals the column 
should be separated in.}

\item{open.left.closed.right}{Logical variable specifying 
whether the output should be in the format 
[open left, closed right) or (closed left, open right]}

\item{method}{The method used to generate the class 
intervals.}

\item{intervals}{If method="specified" this needs to be 
provided as a vector of numeric values. The length of the 
vector needs to be num.intervals-1 and the minimum and 
maximum needs to be within range of the minimum and maximum 
of the selected column.}

\item{labels}{Optional labels for the intervals. By default 
the range of the labels will be used.}
}
\value{
The same data.frame as dafr except that the class 
intervals are added as additional column or the the unchanged 
dafr data.frame is returned if the input is wrong. Warnings 
are provided in this case.
}
\description{
Form class intervals from a column specified by column 
name or column index from the a data.frame.
}
\note{
This is a wrapper for the cut function. See \code{?cut}.
}
\author{
Christoph Knapp
}

