% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_categorical.R
\name{get.missing.categorical}
\alias{get.missing.categorical}
\title{Convert missing values to categorical variables}
\usage{
get.missing.categorical(.data, vars)
}
\arguments{
\item{.data}{a dataframe with the columns to convert 
its missing values into categorical}

\item{vars}{a character vector of the variables in \code{.data} 
for conversion of missing values to categorical}
}
\value{
original dataframe containing new columns of the converted variables 
for the missing values
with tidyverse code attached
}
\description{
Turn <NA>'s into a "missing" character; 
hence numeric variables will be converted to categorical variables
with any numeric values will be converted to "observed",
and returns the result along with tidyverse code used to generate it.
}
\examples{
converted <- get.missing.categorical(iris, vars = c("Species", "Sepal.Length"))
code(converted)
head(converted)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin
}
