% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzdf.R
\name{inzdf}
\alias{inzdf}
\alias{inzdf.data.frame}
\alias{inzdf.SQLiteConnection}
\title{iNZight data frame object}
\usage{
inzdf(x, name, ...)

\method{inzdf}{data.frame}(x, name = deparse(substitute(x)), ...)

\method{inzdf}{SQLiteConnection}(x, name = deparse(substitute(x)), ...)
}
\arguments{
\item{x}{a data.frame or db connection}

\item{name}{the name of the data}

\item{...}{additional arguments passed to methods}
}
\value{
an \code{inzdf} object
}
\description{
This object allows the data to be either a standard R \code{data.frame} or
a connection to a database.
}
