% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{iNZexport}
\alias{iNZexport}
\title{iNZight Metadata Writer}
\usage{
iNZexport(obj)
}
\arguments{
\item{obj}{The object whose metadata has to be written.}
}
\description{
Takes an object (list or a dataframe) and converts it to metadata.
Stores it in a text file "output.txt".
}
\details{
\code{iNZexport} Creates a new file "output.txt" with metdata for the object.
}
\author{
Akshay Gupta
}
