\name{iNZightTools-package}
\alias{iNZightTools-package}
\alias{iNZightTools}
\docType{package}
\title{
Tools for iNZight
}
\description{
This package contains wrapper functions to be used by the iNZight user
interface, and uses other iNZight packages (such as iNZightRegression,
...). It is not user-accessible.
}
\details{
\tabular{ll}{
Package: \tab iNZightTools\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2013-11-29\cr
License: \tab GPL-2\cr
}
}
\author{
Tom Elliott

Maintainer: Tom Elliott <tell029@aucklanduni.ac.nz>
}

