% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightTools.R
\docType{package}
\name{iNZightTools}
\alias{iNZightTools}
\alias{iNZightTools-package}
\title{Tools for data processing with iNZight}
\description{
The iNZightTools package contains a suite of helper functions
for iNZight, mostly to make GUI developement easier to provide
some type of consistency across desktop and shiny versions.
}
\author{
Tom Elliott et al.
}
\seealso{
\code{iNZight}
}
\keyword{iNZight}

