% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{iNZread}
\alias{iNZread}
\alias{iNZread.default}
\alias{iNZread.delim}
\title{iNZight Import Data}
\usage{
iNZread(obj, ...)

\method{iNZread}{default}(path, extension = tools::file_ext(path),
  preview = FALSE, ...)

\method{iNZread}{delim}(obj, ..., number.of.rows = Inf, col.names = TRUE,
  col.types = NULL, encoding.style = "UTF8", delim = attr(obj, "delim"),
  date.names = "en", time.format = "\%AT", time.zone = Sys.timezone(),
  date.format = "\%Y-\%m-\%d",
  decimal.mark = (Sys.localeconv())["decimal_point"],
  grouping.mark = (Sys.localeconv())["grouping"])
}
\arguments{
\item{...}{additional arguments}

\item{path}{A string. Specifies the location of the file to be read.}

\item{extension}{A string. Specifies the extension of the file.}

\item{preview}{A logical scalar. Should the whole data file be read
or only a part of the file be read (100 lines by default where
possible.).}

\item{number.of.rows}{number of rows to read}
}
\value{
A dataframe of length 100 or less if the file extension
is "csv" or "txt" (if partial file reading is possible) or a
dataframe of length equal to the file.
}
\description{
Reads the file and returns a dataframe according to the arguments
passed.
}
\details{
\code{iNZread} returns a dataframe by converting the data in the
file passed based on the arguments included while passing the file.
}
\examples{
iNZread("\\user\\Akshay\\datafile.txt")
iNZread("\\user\\Akshay\\datafile.txt", "csv")
iNZread("\\user\\Akshay\\datafile.csv", "csv", TRUE)
iNZread("\\user\\Akshay\\datafile.csv", col.names = FALSE)

}
\author{
Akshay Gupta
}

