% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_dt}
\alias{is_dt}
\title{Is datetime check}
\usage{
is_dt(x)
}
\arguments{
\item{x}{the variable to check}
}
\value{
logical, \code{TRUE} if the variable is a datetime
}
\description{
This function checks if a variable a date/time/datetime
}
\author{
Tom Elliott
}
