% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_factor}
\alias{is_factor}
\title{Is factor check}
\usage{
is_factor(x)
}
\arguments{
\item{x}{the variable to check}
}
\value{
logical, \code{TRUE} if the variable is numeric
}
\description{
This function checks if a variable a factor.
}
\author{
Tom Elliott
}
