% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_numeric}
\alias{is_numeric}
\title{Is numeric check}
\usage{
is_numeric(x)
}
\arguments{
\item{x}{the variable to check}
}
\value{
logical, \code{TRUE} if the variable is numeric
}
\description{
This function checks if a variable is numeric,
or could be considered one.
For example, dates and times can be treated as numeric,
so return \code{TRUE}.
}
\author{
Tom Elliott
}
