% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_survey}
\alias{is_survey}
\title{Check if object is a survey object (either standard or replicate design)}
\usage{
is_survey(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
logical
}
\description{
Check if object is a survey object (either standard or replicate design)
}
\author{
Tom Elliott
}
