% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_svydesign}
\alias{is_svydesign}
\title{Check if object is a survey object (created by svydesign())}
\usage{
is_svydesign(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
logical
}
\description{
Check if object is a survey object (created by svydesign())
}
\author{
Tom Elliott
}
