% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_svyrep}
\alias{is_svyrep}
\title{Check if object is a replicate survey object (created by svrepdesign())}
\usage{
is_svyrep(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
logical
}
\description{
Check if object is a replicate survey object (created by svrepdesign())
}
\author{
Tom Elliott
}
