% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{isColumnTypesCorrect}
\alias{isColumnTypesCorrect}
\title{Checks if the entered column types are allowed or not.}
\usage{
isColumnTypesCorrect(col.types)
}
\arguments{
\item{col.types}{a vector of datatypes of columns.}
}
\value{
a logical scalar.
}
\description{
\code{isColumnTypesCorrect} returns a logical scalar to tell if the
column types entered are valid. Returns TRUE if they are valid.
Returns FALSE if not valid.
}

