% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{isFullFile}
\alias{isFullFile}
\title{Checks if the complete file was read or not.}
\usage{
isFullFile(user.data.frame)
}
\arguments{
\item{user.data.frame}{the dataframe returned after reading the file}
}
\value{
A logical scalar
}
\description{
\code{isFullFile} returns a logical scalar after checking 
the attributes of the dataframe sent as an argument. 
Returns TRUE if it is the complete file. Returns FALSE if 
it is a partial file read.
}

