% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{isMetadataAvailable}
\alias{isMetadataAvailable}
\title{Reads a file and checks if metadata is available.}
\usage{
isMetadataAvailable(path)
}
\arguments{
\item{path}{A string. Specifies the location of the file to be read.}
}
\value{
A logical vector of length 1.
}
\description{
\code{isMetadataAvailable} Returns TRUE if metadata is available.
}

