% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{isPreview}
\alias{isPreview}
\title{Checks if the complete file was read or not.}
\usage{
isPreview(user.data.frame)
}
\arguments{
\item{user.data.frame}{the dataframe returned after reading the file}
}
\value{
A logical scalar
}
\description{
\code{isPreview} returns a logical scalar after checking
the attributes of the dataframe sent as an argument.
Returns TRUE if it was read as a preview.
}
