% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{join_data}
\alias{join_data}
\title{Join data with another dataset}
\usage{
join_data(
  data_l,
  data_r,
  by = NULL,
  how = c("inner", "left", "right", "full", "anti", "semi"),
  suffix_l = ".x",
  suffix_r = ".y"
)
}
\arguments{
\item{data_l}{original data}

\item{data_r}{imported dataset}

\item{by}{a character vector of variables to join by}

\item{how}{the method used to join the datasets}

\item{suffix_l}{suffix for the original dataset (ignored for filter-joins)}

\item{suffix_r}{suffix for the imported dataset (ignored for filter-joins)}
}
\value{
joined dataset
}
\description{
Join data with another dataset
}
\seealso{
\code{\link{code}}, \code{\link[dplyr]{mutate-joins}},
         \code{\link[dplyr]{filter-joins}}
}
\author{
Zhaoming Su
}
