% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{joindata}
\alias{joindata}
\title{Join data with another dataset}
\usage{
joindata(
  .data,
  imported_data,
  origin_join_col,
  import_join_col,
  join_method,
  left,
  right
)
}
\arguments{
\item{.data}{Original data}

\item{imported_data}{Imported dataset}

\item{origin_join_col}{column selected from the original data}

\item{import_join_col}{column selected from the imported dataset}

\item{join_method}{function used to join the two datasets}

\item{left}{suffix name assigned to the original dataset}

\item{right}{suffix name assigned to the imported dataset}
}
\value{
joined dataset
}
\description{
Join data with another dataset
}
