% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_linked.R
\name{load_linked}
\alias{load_linked}
\title{Import linked data into an \code{inzdf} object}
\usage{
load_linked(x, schema, con, name = deparse(substitute(con)), ...)
}
\arguments{
\item{x}{a linked specification file or vector of data set paths}

\item{schema}{a list describing the schema/relationships between the files}

\item{con}{a database connection to load the linked data into}

\item{name}{the name of the data set collection}

\item{...}{additional arguments passed to data reading function \code{smart_read()}}
}
\value{
an \code{inzdf} object
}
\description{
Import linked data into an \code{inzdf} object
}
