% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{load_rda}
\alias{load_rda}
\title{Load object(s) from an Rdata file}
\usage{
load_rda(file)
}
\arguments{
\item{file}{path to an rdata file}
}
\value{
list of data frames, plus code
}
\description{
Load object(s) from an Rdata file
}
\seealso{
\code{\link{save_rda}}
}
\author{
Tom Elliott
}
