% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_names}
\alias{make_names}
\title{Make unique variable names}
\usage{
make_names(new, existing = character())
}
\arguments{
\item{new}{a vector of proposed new variable names}

\item{existing}{a vector of existing variable names}
}
\value{
a vector of unique variable names
}
\description{
Helper function to create new variable names that are unique
given a set of existing names (in a data set, for example).
If a variable name already exists, a number will be appended.
}
\examples{
make_names(c("var_x", "var_y"), c("var_x", "var_z"))

}
\author{
Tom Elliott
}
