% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{makeLocale}
\alias{makeLocale}
\title{Creates a locale that cane be used with functions from readr package.}
\usage{
makeLocale(date.names, date.format, time.format, decimal.mark, grouping.mark,
  time.zone, encoding.style)
}
\arguments{
\item{date.names}{The language used in the file to specift names 
of months.}

\item{date.format}{the format of date in the file.}

\item{time.format}{the format of time in the file.}

\item{decimal.mark}{the symbol used as the decimal mark in the file.}

\item{grouping.mark}{the symbol used as the grouping mark in the file.}

\item{time.zone}{the timezone used while writing dates/time 
in the file.}

\item{encoding.style}{The encoding style used to make the file.}
}
\value{
Zn object of class locale.
}
\description{
\code{makeLocale} returns an object of class locale that has 
been created using the arguments passed. The class 
locale is defined in the package readr.
}

