% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_as_date}
\alias{mutate_as_date}
\title{Convert variables to dates}
\usage{
mutate_as_date(data, vars, names = NULL)
}
\arguments{
\item{data}{a dataframe with the variables to convert}

\item{vars}{a character vector of column names to convert}

\item{names}{a character vector of names for the created variables}
}
\value{
original dataframe containing new columns of the
        converted variables with tidyverse code attached
}
\description{
Convert variables to dates
}
\seealso{
\code{\link{code}}
}
\author{
Stephen Su
}
