% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdevice.R
\name{newdevice}
\alias{newdevice}
\title{Open a New Graphics Device}
\usage{
newdevice(width = 7, height = 7, ...)
}
\arguments{
\item{width}{the width (in inches) of the new device}

\item{height}{the height (in inches) of the new device}

\item{...}{additional arguments passed to the new device function}
}
\description{
Opens a new graphics device
}
\details{
Depending on the system, difference devices are better.
}
\author{
Tom Elliott
}

