% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamod_numeric.R
\name{numToCat}
\alias{numToCat}
\title{Convert to Categorical}
\usage{
numToCat(.dataset, var, name = sprintf("\%s.cat", var))
}
\arguments{
\item{.dataset}{the dataset to use}

\item{var}{character, name of the numeric variable}

\item{name}{name of the created variable (defaults to {original}.cat)}
}
\value{
a dataframe with the new variable, along with the code used to generate it
}
\description{
Convert a numeric variable to Categorical (factor)
}
\author{
Tom Elliott
}
