% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Anti value matching}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector of values to be matched}

\item{table}{vector of values to match against}
}
\value{
A logical vector of same length as 'x', indicating if each
element does \strong{not} exist in the table.
}
\description{
Anti value matching
}
