% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%||\%}
\alias{\%||\%}
\title{NULL or operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{an object, potentially NULL}

\item{b}{an object}
}
\value{
a if a is not NULL, otherwise b
}
\description{
NULL or operator
}
