% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.data.R
\name{order.overwrite}
\alias{order.overwrite}
\title{Order vectors}
\usage{
order.overwrite(z, na.last = TRUE, decreasing = FALSE)
}
\arguments{
\item{z}{a sequence of numeric, complex, character, or logical vectors,
all of the same length, or a classed R object.}

\item{na.last}{for controlling the treatment of NAs. If \code{TRUE},
missing values in the data are put last; otherwise they are put first.
If \code{NA}, they are removed (see \code{Note}).}

\item{decreasing}{logical; if \code{TRUE}, the order will be decreasing,
otherwise increasing.}
}
\value{
a numeric vector of orderings
}
\description{
The iNZight version of the order function which lets you pass
in a list of vectors to order instead of the \code{...} argument.
It is shortened and might therefore not be as stable as the original
order function.
}
\note{
This function is only called in \code{sort.data} but needs to be
available to \code{sort.data}.
}
\author{
Christoph Knapp
}
