% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_survey.R
\name{print.inzsvyspec}
\alias{print.inzsvyspec}
\title{Print iNZight Survey Spec}
\usage{
\method{print}{inzsvyspec}(x, ...)
}
\arguments{
\item{x}{a \code{inzsvyspec} object}

\item{...}{additional arguments, ignored}
}
\description{
Print iNZight Survey Spec
}
\author{
Tom Elliott
}
