% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_writing.R
\name{print_code}
\alias{print_code}
\title{Tidy-printing of the code attached to an object}
\usage{
print_code(x, ...)
}
\arguments{
\item{x}{a dataframe with code attached}

\item{...}{additional arguments passed to tidy_all_code()}
}
\value{
Called for side-effect of printing code to the console.
}
\description{
Tidy-printing of the code attached to an object
}
\examples{
iris_agg <- aggregateData(iris, vars = "Species", summaries = "mean")
print_code(iris_agg)
}
