% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dict.R
\name{read_dictionary}
\alias{read_dictionary}
\title{Read a file as a data dictionary}
\usage{
read_dictionary(file, id, name, description, ...)
}
\arguments{
\item{file}{path to a file}

\item{id}{name of the column containing the ID, if missing will be guessed}

\item{name}{name of the column containing the human-readable variable name}

\item{description}{name of the column containing the variable description}
}
\description{
Read a file as a data dictionary
}
\section{Columns}{

The following columns are understood:
\itemize{
\item units - these will be used to label axes / legends
}
}

