% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_text.R
\name{read_text}
\alias{read_text}
\title{Read text as data}
\usage{
read_text(txt, delim = "\\t", ...)
}
\arguments{
\item{txt}{character string}

\item{delim}{the delimiter to use, passed to `readr::read_delim()`}

\item{...}{additional arguments passed to `readr::read_delim()`}
}
\value{
data.frame
}
\description{
The text can also be the value `"clipboard"` which
will use `readr::clipboard()`.
}
\author{
Tom Elliott
}
