% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_read.R, R/read_metadata.R
\name{readMetadata}
\alias{readMetadata}
\alias{readMetadata}
\title{Read CSV with iNZight metadata}
\usage{
readMetadata(path)

readMetadata(path)
}
\arguments{
\item{path}{A string. Specifies the location of the file to be read.}

\item{file}{the plain text file with metadata}
}
\value{
a data frame

A list.
}
\description{
This function will read a CSV file with iNZight metadata in the header.
This allows plain text CSV files to be supplied with additional comments
that describe the structure of the data to make import and data handling easier.

Reads a file and checks for metadata if available.
}
\details{
The main example is to define factor levels for an integer variable in large data sets.

\code{readMetadata} Returns the metadata if available in the file as a list.
}
\author{
Tom Elliott

Akshay Gupta
}
