% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_read.R
\name{readMetadata2}
\alias{readMetadata2}
\title{Read CSV with iNZight metadata}
\usage{
readMetadata2(file)
}
\arguments{
\item{file}{the plain text file with metadata}
}
\value{
a data frame
}
\description{
This function will read a CSV file with iNZight metadata in the header.
This allows plain text CSV files to be supplied with additional comments
that describe the structure of the data to make import and data handling easier.
}
\details{
The main example is to define factor levels for an integer variable in large data sets.
}
\author{
Tom Elliott
}
