% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{remDoubleSpaces}
\alias{remDoubleSpaces}
\title{Remove multiple spaces and replace with a single space.}
\usage{
remDoubleSpaces(string)
}
\arguments{
\item{string}{A string. A line in the metadata.}
}
\value{
A string.
}
\description{
\code{remDoubleSpaces} Returns a formatted string with 
multiple spacces substituted with a single space.
}

