% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{remove_rows}
\alias{remove_rows}
\title{Remove rows from data by row numbers}
\usage{
remove_rows(data, rows)
}
\arguments{
\item{data}{a dataframe or a survey design object to filter}

\item{rows}{a numeric vector of row numbers to slice off}
}
\value{
filtered dataframe with tidyverse code attached
}
\description{
Filter a dataframe by slicing off specified rows
and returns the result along with tidyverse code used to generate it.
}
\examples{
data <- remove_rows(iris, rows = c(1, 4, 5))
cat(code(data))
head(data)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin, Zhaoming Su
}
