% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{removeUnneededSpaces}
\alias{removeUnneededSpaces}
\title{Removes unneeded spaces like double spaces or spaces around operators.}
\usage{
removeUnneededSpaces(string)
}
\arguments{
\item{string}{A string. A line in the metadata.}
}
\value{
A string.
}
\description{
\code{removeUnneededSpaces} Returns a formatted string with 
unneeded spaces removed.
}

