% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{remSpacesAroundOperators}
\alias{remSpacesAroundOperators}
\title{iNZight Import Data}
\usage{
remSpacesAroundOperators(string)
}
\arguments{
\item{path}{A string. Specifies the location of the file to be read.}

\item{...}{additional arguments}
}
\value{
A dataframe.
}
\description{
Reads the file and returns a dataframe according to the arguments
passed.
}
\details{
\code{.remSpacesAroundOperators} returns a dataframe by converting the data in the
file passed based on the arguments included while passing the file.
}
\author{
Akshay Gupta
}

