% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_levels.R
\name{rename.levels}
\alias{rename.levels}
\title{Rename the levels of a categorical variable}
\usage{
rename.levels(.data, var, to_be_renamed)
}
\arguments{
\item{.data}{a dataframe with the column to be renamed}

\item{var}{a character of the categorical variable to rename}

\item{to_be_renamed}{a list of the old level name assigned to the new level name;
ie. list('new level name' = 'old level name')}
}
\value{
original dataframe containing a new column of the renamed categorical variable with tidyverse code attached
}
\description{
Rename the levels of a categorical variables, and returns the result
along with tidyverse code used to generate it.
}
\examples{
renamed <- rename.levels(iris, var = "Species", to_be_renamed = 
list(set = "setosa", ver = "versicolor"))
code(renamed)
head(renamed)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin
}
