% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{rename_vars}
\alias{rename_vars}
\title{Rename column names}
\usage{
rename_vars(data, from, to)
}
\arguments{
\item{data}{a dataframe with columns to rename}

\item{from}{a character vector the variable names as-is}

\item{to}{a character vector the variable names to-be}
}
\value{
original dataframe containing new columns of the renamed columns
       with tidyverse code attached
}
\description{
Rename columns of a dataset with desired names
}
\examples{
renamed <- rename_vars(iris,
    from = names(iris),
    to = tolower(gsub("\\\\.", "_", names(iris)))
)
cat(code(renamed))
head(renamed)

}
\seealso{
\code{\link{code}}
}
\author{
Stephen Su
}
