% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_vars.R
\name{renameVars}
\alias{renameVars}
\title{Rename column names}
\usage{
renameVars(.data, to_be_renamed_list)
}
\arguments{
\item{.data}{a dataframe with columns to rename}

\item{to_be_renamed_list}{a list of the new column names assigned
to the old column names
ie. list('old column names' = 'new column names')}
}
\value{
original dataframe containing new columns of the renamed columns
with tidyverse code attached
}
\description{
Rename column names
and returns the result along with tidyverse code used to generate it.
}
\examples{
renamed <- renameVars(iris,
    to_be_renamed_list = list(Species = "Type", Petal.Width = "P.W"))
cat(code(renamed))
head(renamed)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin
}
