% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_levels.R
\name{reorder.levels}
\alias{reorder.levels}
\title{This function changes the order of levels in the data.}
\usage{
\method{reorder}{levels}(dafr, column, levels.new)
}
\arguments{
\item{dafr}{The data to be changed}

\item{column}{the factor column where the order of 
levels should be changed.}

\item{levels.new}{A vector of all levels in the column 
specified by column in the order they should be 
ordered.}
}
\value{
A data.frame with the levels of one column 
reordered
}
\description{
This function changes the order of levels in the data.
}
\author{
Christoph Knapp
}

