% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{replace_na}
\alias{replace_na}
\alias{replace_na.factor}
\title{Replace NAs with specified values}
\usage{
\method{replace_na}{factor}(data, replace = "<NA>", ...)
}
\arguments{
\item{data}{A data frame or vector.}

\item{replace}{If \code{data} is a factor, the default value
will be \code{"<NA>"}.}

\item{...}{Additional arguments for methods. Currently unused.}
}
\description{
Replace NAs with specified values

Make \code{\link[tidyr]{replace_na}} also work for factors.
}
