% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{replaceComma}
\alias{replaceComma}
\title{Replaces comma inside factor's values with '?'.}
\usage{
replaceComma(string)
}
\arguments{
\item{string}{A string. A line in the metadata.}
}
\value{
A string.
}
\description{
\code{replaceComma} Returns a string with '?' instead of ',' inside 
brackets for factor values.
}

