% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{reshape_data}
\alias{reshape_data}
\title{Reshaping dataset from wide to long or from long to wide}
\usage{
reshape_data(
  data,
  data_to = c("long", "wide"),
  cols,
  names_to = "name",
  values_to = "value",
  names_from = "name",
  values_from = "value"
)
}
\arguments{
\item{data}{a dataset to reshape}

\item{data_to}{whether the target dataset is \code{long} or \code{wide}}

\item{cols}{columns to gather together (for wide to long)}

\item{names_to}{name for new column containing old names (for wide to long)}

\item{values_to}{name for new column containing old values (for wide to long)}

\item{names_from}{column to spread out (for long to wide)}

\item{values_from}{values to be put in the spread columns (for long to wide)}
}
\value{
reshaped dataset
}
\description{
Reshaping dataset from wide to long or from long to wide
}
\author{
Zhaoming Su
}
