% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{reshape_data_long_to_wide}
\alias{reshape_data_long_to_wide}
\title{Reshape long to wide}
\usage{
reshape_data_long_to_wide(.data, col1, col2)
}
\arguments{
\item{.data}{a dataframe or tibble to be reshaped}

\item{col1}{the column containing the 'key'}

\item{col2}{the column containing the values}
}
\value{
a dataframe or tibble
}
\description{
Reshape a data set from long form (observations in rows),
to wide form (observations in columns).
}
\author{
Yiwen He
}
