% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{reshape_data_wide_to_long}
\alias{reshape_data_wide_to_long}
\title{Reshape wide to long}
\usage{
reshape_data_wide_to_long(.data, cols, colname, value)
}
\arguments{
\item{.data}{a dataframe or tibble to be reshaped}

\item{cols}{the columns containing values}

\item{colname}{the name of the 'key' column in the new data set}

\item{value}{the name of the 'value' column in the new data set}
}
\value{
a dataframe or tibble
}
\description{
Reshape a data set from wide form (observations in columns)
to long form (observations in rows)
}
\author{
Yiwen He
}
