% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{returnWords}
\alias{returnWords}
\title{Converts the string into a list of words.}
\usage{
returnWords(string)
}
\arguments{
\item{string}{A string. A line in the metadata.}
}
\value{
A list.
}
\description{
\code{returnWords} Returns a list of words present in the line.
}

