% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{rows_remove}
\alias{rows_remove}
\title{Remove rows from data by row numbers}
\usage{
rows_remove(data, rows)
}
\arguments{
\item{data}{a dataframe or a survey design object to filter}

\item{rows}{a numeric vector of row numbers to slice off}
}
\value{
filtered dataframe with tidyverse code attached
}
\description{
Filter a dataframe by slicing off specified rows
and returns the result along with tidyverse code used to generate it.
}
\examples{
filtered <- rows_remove(iris, rows = c(1, 4, 5))
code(filtered)
head(filtered)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin, Stephen Su
}
