% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample.data}
\alias{sample.data}
\title{Takes a sample of rows from a data.frame}
\usage{
sample.data(df, sampleSize, numSample = 1, bootstrap = F)
}
\arguments{
\item{df}{A data.frame the sample is taken from.}

\item{sampleSize}{The size of the samples to be taken.}

\item{numSample}{The number of samples to be taken.}

\item{bootstrap}{TRUE if samples with replacement is 
desired, FALSE if no replacement.}
}
\value{
A data.frame with the samples merged together. 
An additional column is added where the sampling 
iteration is stored.
}
\description{
This function samples rows from a data.frame with or 
without replacment.
}
\author{
Christoph Knapp
}

