% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{save_rda}
\alias{save_rda}
\title{Save an object with, optionally, a (valid) name}
\usage{
save_rda(data, file, name)
}
\arguments{
\item{data}{the data frame to save}

\item{file}{where to save it}

\item{name}{optional, the name the data will have in the rda file}
}
\value{
logical, should be TRUE, along with code for the save
}
\description{
Save an object with, optionally, a (valid) name
}
\seealso{
\code{\link{load_rda}}
}
\author{
Tom Elliott
}
