% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_name.R
\name{search.name}
\alias{search.name}
\title{Searches a list recursivly for a name and
returns the value associated with the name}
\usage{
search.name(list.search, search.name = NULL)
}
\arguments{
\item{list.search}{A list to search}

\item{search.name}{A name to be found in a list. 
If null, the whole simplified list is returned.}
}
\value{
The value found in the list associated 
with search.name, NULL if not there.
}
\description{
Searches a list recursivly for a name and
returns the value associated with the name
}
\details{
This function is used to search the 
everchanging plot output from iNZightPlot. 
It can handle 4 classes found in the plot 
output (list, inzplotoutput, inzgrid,
inzpar.list) which can be treated as list 
type object. Some of the objects retrieved 
might also be a list. This happens when the 
name to search for is found several times in 
the nested list structure of the plot output.
}
\author{
Christoph Knapp
}

