% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vars.R
\name{select_vars}
\alias{select_vars}
\title{Select variables from a dataset}
\usage{
select_vars(data, keep)
}
\arguments{
\item{data}{the dataset}

\item{keep}{vector of variable names to keep}
}
\value{
a data frame with tidyverse code attribute
}
\description{
Select a (reordered) subset of variables from a subset.
}
\examples{
select_vars(iris, c("Sepal.Length", "Species", "Sepal.Width"))
}
\author{
Tom Elliott, Zhaoming Su
}
