% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_col.R
\name{separate}
\alias{separate}
\title{Separate columns}
\usage{
separate(.data, col, left, right, sep, check)
}
\arguments{
\item{.data}{dataset}

\item{col}{column to be separated}

\item{left}{name for the separated left column}

\item{right}{name for the separated right column}

\item{sep}{separator used to separate columns}

\item{check}{method of separating}
}
\value{
separated dataset
}
\description{
Separate columns
}
\author{
Yiwen He, Tom Elliott
}
