% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_var.R
\name{separate_var}
\alias{separate_var}
\title{Separate columns}
\usage{
separate_var(
  data,
  var,
  left,
  right,
  by,
  into = c("cols", "rows"),
  method = c("delim", "position")
)
}
\arguments{
\item{data}{dataset}

\item{var}{name of variable to be separated}

\item{left}{name for the separated left column (for \code{into = "cols"})}

\item{right}{name for the separated right column (for \code{into = "cols"})}

\item{by}{a string as delimiter between values (\code{delim} method) or an
integer as number of characters to split by (\code{position} method)}

\item{into}{whether to split into new rows or columns}

\item{method}{method of separation}
}
\value{
Separated dataset
}
\description{
Separate columns
}
\author{
Stephen Su
}
