% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{sheets}
\alias{sheets}
\title{List of available sheets from a file}
\usage{
sheets(x)
}
\arguments{
\item{x}{a dataframe from \code{smart_read}}
}
\value{
vector of sheet names, or NULL
}
\description{
List of available sheets from a file
}
\author{
Tom Elliott
}
