% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{solveFirstWord}
\alias{solveFirstWord}
\title{Solves the first word in the metadata line.}
\usage{
solveFirstWord(first.word)
}
\arguments{
\item{first.word}{A string. First word/expression/phrase in the metadata line.}
}
\value{
A string or a logical vector of length 1.
}
\description{
\code{solveSecondWord} Returns variable name if correct metadata is present. 
Returns FALSE otherwise.
}

