% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{solveForValues}
\alias{solveForValues}
\alias{solveForValues.default}
\alias{solveForValues.with.levels}
\alias{solveForValues.without.levels}
\title{Extracts the levels and labels to be used when converting a column to factor.}
\usage{
solveForValues(obj)

\method{solveForValues}{default}(obj)

\method{solveForValues}{with.levels}(obj)

\method{solveForValues}{without.levels}(obj)
}
\arguments{
\item{obj}{A string.}
}
\value{
A character vector.
}
\description{
\code{readMetadata} Returns a vector with label and levels.
}

