% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{solveSecondWord}
\alias{solveSecondWord}
\alias{solveSecondWord.default}
\alias{solveSecondWord.iNZmeta.date}
\alias{solveSecondWord.iNZmeta.factor}
\alias{solveSecondWord.iNZmeta.numeric}
\title{Solves the second word in the metadata line which can be of the 
form numeric,character, factor["x=xyz"]. It adds the details of the metadata in a list.}
\usage{
solveSecondWord(second.word, variable.specs)

\method{solveSecondWord}{default}(second.word, variable.specs)

\method{solveSecondWord}{iNZmeta.factor}(second.word, variable.specs)

\method{solveSecondWord}{iNZmeta.numeric}(second.word, variable.specs)

\method{solveSecondWord}{iNZmeta.date}(second.word, variable.specs)
}
\arguments{
\item{second.word}{A string. Second word/expression/phrase in the metadata line.}
}
\value{
A list.
}
\description{
\code{solveSecondWord} Returns a list updated with the datatype of the column 
and its values as an attribute if available.
}

