% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.data.R
\name{sort.data}
\alias{sort.data}
\title{Sorts the df data frame after the input variables}
\usage{
\method{sort}{data}(vars, sort.type, df)
}
\arguments{
\item{vars}{The column names in the order the df data.frame
should be sorted.}

\item{sort.type}{A logical vector of the same length as 
vars. If the element in the vector is TRUE the corresponding 
element in vars will be sorted in increasing order.}

\item{df}{The data.frame or matrix to sort.}
}
\value{
An ordered data.frame
}
\description{
Warnings will be given if the colnames in vars will not 
match the column in df. This is a wrapper function for 
sort. See \code{?sort} for more information.
}
\author{
Christoph Knapp
}

