% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_variables.R
\name{stackVars}
\alias{stackVars}
\title{Stack variables}
\usage{
stackVars(.data, vars, key = "stack.variable", value = "stack.value")
}
\arguments{
\item{.data}{a dataframe to stack}

\item{vars}{a character vector of variables to stack}

\item{key}{name of the new column for the stacked variables.
"stack.variable" by default}

\item{value}{name of the new column for the stacked values of the stacked.
"stack.value" by default}
}
\value{
stacked dataframe with tidyverse code attached
}
\description{
Collapse columns by converting from a wide to a long format
and returns the result along with tidyverse code used to generate it.
}
\examples{
stacked <- stackVars(iris, vars = c("Species", "Sepal.Width"),
    key = "Variable", value = "Value")
cat(code(stacked))
head(stacked)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin
}
