% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_var.R
\name{standardizeVars}
\alias{standardizeVars}
\title{Transform data of a numeric variable}
\usage{
standardizeVars(.data, vars)
}
\arguments{
\item{.data}{a dataframe with the columns to standarize}

\item{vars}{a character vector of the numeric variables in \code{.data} 
to standardize}
}
\value{
the original dataframe containing new columns of the standardized variables with tidyverse code attached
}
\description{
Transform the values of a numeric variable by applying a mathematical function
}
\examples{
standardized <- standardizeVars(iris, var = c("Sepal.Width", "Petal.Width"))
code(standardized)
head(standardized)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin
}
