% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyspec.R
\name{import_survey}
\alias{import_survey}
\alias{make_survey}
\alias{as_survey}
\alias{as_survey_spec}
\alias{write_spec}
\title{Survey Specification helper functions}
\usage{
import_survey(...)

make_survey(...)

as_survey(...)

as_survey_spec(...)

write_spec(...)
}
\arguments{
\item{...}{arguments passed to 'surveyspec' function}
}
\value{
see 'surveyspec::import_survey'

see 'surveyspec::make_survey'

see 'surveyspec::as_survey'

see 'surveyspec::as_survey_spec'

see 'surveyspec::write_spec'
}
\description{
These functions are just wrappers around the functions in 'surveyspec' so existing code continues to work.
}
\section{Functions}{
\itemize{
\item \code{import_survey()}: Import a survey object

\item \code{make_survey()}: Make a survey

\item \code{as_survey()}: as survey method

\item \code{as_survey_spec()}: as survey spec method

\item \code{write_spec()}: write a survey specification

}}
