% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycode.R
\name{tidy_all_code}
\alias{tidy_all_code}
\title{iNZight Tidy Code}
\usage{
tidy_all_code(messy_code, width, indent, outfile, incl_library = TRUE)
}
\arguments{
\item{messy_code}{file name of the file containing messy code}

\item{width}{the width of a line}

\item{indent}{how many spaces for one indent}

\item{outfile}{the file name of the file containing formatted code}

\item{incl_library}{logical, if true, the output code will contain library name}
}
\value{
a file with formatted code
}
\description{
Tidy code with correct indents and limit the code to the specific width
}
\author{
Lushi Cai
}
