% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamod_numeric.R
\name{transform}
\alias{transform}
\title{Transform Numeric Variable}
\usage{
transform(.dataset, var, fn, name = sprintf("\%s.\%s", fn, var))
}
\arguments{
\item{.dataset}{the dataset to use}

\item{var}{character, name of the numeric variable}

\item{fn}{character, name of the function to call}

\item{name}{name of the created variable (defaults to {original}.{fn})}
}
\value{
a dataframe with the new variable(, along with the code used to generate it)
}
\description{
Transform a numeric variable
}
\author{
Tom Elliott
}
