% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_cols.R
\name{unite}
\alias{unite}
\title{Unite columns in a dataset}
\usage{
unite(.data, name, col, sep)
}
\arguments{
\item{.data}{dataset}

\item{name}{name for the new united column}

\item{col}{a vector of column names}

\item{sep}{separator used in between the united columns}
}
\value{
united dataset
}
\description{
Unite columns in a dataset
}
\author{
Yiwen He
}
