% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{url_to_temp}
\alias{url_to_temp}
\title{Download URL to temp file}
\usage{
url_to_temp(url)
}
\arguments{
\item{url}{where the file lives on the internet}
}
\value{
the location of a (temporary) file location
}
\description{
Download URL to temp file
}
\author{
Tom Elliott
}
