% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_dataset.R
\name{validation_details}
\alias{validation_details}
\title{Details of Validation Rule Results}
\usage{
validation_details(cf, v, var, id.var, df)
}
\arguments{
\item{cf}{Confrontation object from \code{validate::confront()}}

\item{v}{Validator that generated \code{cf}}

\item{var}{Rule name to give details about}

\item{id.var}{Variable name denoting a unique identifier
for each observation}

\item{df}{The dataset that was confronted}
}
\value{
A character vector giving each line of the summary detail text
}
\description{
Generates the more detailed text required for the details section in
\code{iNZValidateWin}.
}
\author{
Daniel Barnett
}
