% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_dataset.R
\name{validation_summary}
\alias{validation_summary}
\title{Validation Confrontation Summary}
\usage{
validation_summary(cf)
}
\arguments{
\item{cf}{Confrontation object from \code{validate::confront()}}
}
\value{
A \code{data.frame} with number of tests performed, number of
passes, number of failures, and failure percentage for each validation rule.
}
\description{
Generates a summary of a confrontation which gives basic information about
each validation rule tested.
}
\author{
Daniel Barnett
}
