% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vartype}
\alias{vartype}
\title{Get variable type name}
\usage{
vartype(x)
}
\arguments{
\item{x}{vector to be examined}
}
\value{
character vector of the variable's type
}
\description{
Get variable type name
}
\author{
Tom Elliott
}
