% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vartypes}
\alias{vartypes}
\title{Get all variable types from data object}
\usage{
vartypes(x)
}
\arguments{
\item{x}{data object (data.frame or inzdf)}
}
\value{
a named vector of variable types
}
\description{
Get all variable types from data object
}
