% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate.data}
\alias{aggregate.data}
\title{aggregates the data over selected factor columns}
\usage{
\method{aggregate}{data}(aggregate.over, methods = c("mean", "median", "sum",
  "sd", "IQR", "count"), dafr)
}
\arguments{
\item{aggregate.over}{column names of factor variables 
in data to aggregate over}

\item{methods}{A set of methods which can be used to 
aggregate.}

\item{dafr}{A data.frame containing at least on factor 
column and one numeric column.}
}
\value{
A data.frame with the results of the aggregation.
}
\description{
The dimensions of the df data frame will change so that 
all possible combinations of factors selected will be 
the number of rows and the number of all selected factor 
columns + a column for all methods selected will be the 
number of columns in the return data.frame.
}
\author{
Christoph Knapp
}
