% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_rows.R
\name{append_rows}
\alias{append_rows}
\title{Append rows to a dataset}
\usage{
append_rows(data, new_data, when_added = FALSE)
}
\arguments{
\item{data}{The original dataset to which new rows will be appended.}

\item{new_data}{The dataset containing the new rows.}

\item{when_added}{Logical; indicates whether a \code{.when_added} column
is required.}
}
\value{
A dataset with new rows appended below the original \code{data}.
}
\description{
Append rows to a dataset
}
\author{
Yiwen He, Zhaoming Su
}
