% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dt.R
\name{convert_to_datetime}
\alias{convert_to_datetime}
\title{Convert to datetime}
\usage{
convert_to_datetime(.data, factorname, convname, newname)
}
\arguments{
\item{.data}{dataframe}

\item{factorname}{name of the variable}

\item{convname}{format}

\item{newname}{name of the new column}
}
\value{
dataframe with datetime column
}
\description{
Convert to datetime
}
\author{
Yiwen He
}
