% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{convert_to_datetime}
\alias{convert_to_datetime}
\title{Convert variables to date-time}
\usage{
convert_to_datetime(data, vars, ord = NULL, names = NULL, tz = "")
}
\arguments{
\item{data}{a dataframe with the variables to convert}

\item{vars}{a character vector of column names to convert}

\item{ord}{a character vector of date-time formats}

\item{names}{a character vector of names for the created variables}

\item{tz}{a time zone name (default: local time zone). See
\code{\link[base]{OlsonNames}}}
}
\value{
original dataframe containing new columns of the
        converted variables with tidyverse code attached
}
\description{
Convert variables to date-time
}
\seealso{
\code{\link{code}}
}
\author{
Zhaoming Su
}
