% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missing.R
\name{countMissing}
\alias{countMissing}
\title{Count missing values}
\usage{
countMissing(var, na.rm = FALSE)
}
\arguments{
\item{var}{the vector to sum up the number of missing values}

\item{na.rm}{ignore this}
}
\value{
the number of missing values for that vector
}
\description{
Count missing values
}
\seealso{
\code{\link{aggregateData}}
}
\author{
Owen Jin
}
