% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dt_comp.R
\name{extract_dt_comp}
\alias{extract_dt_comp}
\title{Extract date component from a date-time variable}
\usage{
extract_dt_comp(data, var, comp, name = NULL)
}
\arguments{
\item{data}{The dataframe containing the date-time variable.}

\item{var}{The name of the date-time variable to extract the component.}

\item{comp}{The date component wanted from the variable. See
\code{iNZightTools:::inz_dt_comp} for the full list of components.}

\item{name}{The name of the new column to store the extracted date
component.}
}
\value{
A dataframe with the new date component column.
}
\description{
This function extracts a specific date component from a date-time variable
in a dataframe.
}
\author{
Zhaoming Su
}
