% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extfromdt.R
\name{extract_part}
\alias{extract_part}
\title{Extract part of a datetimes variable}
\usage{
extract_part(.data, varname, part, name)
}
\arguments{
\item{.data}{dataframe}

\item{varname}{name of the variable}

\item{part}{part of the variable wanted}

\item{name}{name of the new column}
}
\value{
dataframe with extracted part column
}
\description{
Extract part of a datetimes variable
}
\author{
Yiwen He
}
