% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.levels.R
\name{get.collapsed.column}
\alias{get.collapsed.column}
\title{Collapses selected levels in factor vector}
\usage{
get.collapsed.column(column, to.collapse)
}
\arguments{
\item{column}{the vector where levels should be 
collapsed into one.}

\item{to.collapse}{Vector of levels to collapse.}
}
\description{
Collapses selected levels in factor vector
}
\note{
Levels in to.collapse which are not in column 
will be ignored.
}
\author{
Christoph Knapp
}

