% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_categorical.R
\name{get.missing.categorical}
\alias{get.missing.categorical}
\title{Converts specified columns into binary factor variables 
which get added to the input data.frame.}
\usage{
get.missing.categorical(dafr, columns)
}
\arguments{
\item{dafr}{The input data.frame.}

\item{columns}{The column names or indexes to be converted.}
}
\value{
A data.frame with the binary columns added.
}
\description{
Converts specified columns into binary factor variables 
which get added to the input data.frame.
}
\author{
Christoph Knapp
}

