% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{iNZread}
\alias{iNZread}
\title{iNZight Import Data}
\usage{
iNZread(path, col.types = NULL, ...)
}
\arguments{
\item{path}{A string. Specifies the location of the file to be read.}

\item{col.types}{A character vector specifying the datatypes of columns.}

\item{...}{additional arguments}
}
\value{
A dataframe.
}
\description{
Reads the file and returns a dataframe according to the arguments
passed.
}
\details{
\code{iNZread} returns a dataframe by converting the data in the
file passed based on the arguments included while passing the file.
}
\author{
Akshay Gupta
}

