% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_types_to_char.R
\name{makeDatatypeChar}
\alias{makeDatatypeChar}
\alias{makeDatatypeChar.character}
\alias{makeDatatypeChar.list}
\title{Converts the column datatypes to a single word or string which can be
passed to readr::read_csv as an argument.}
\usage{
makeDatatypeChar(obj, col.names = NULL)

\method{makeDatatypeChar}{character}(obj)

\method{makeDatatypeChar}{list}(obj, col.names)
}
\arguments{
\item{obj}{A character vector or a list.}
}
\value{
A character vector of length 1.
}
\description{
\code{makeDatatypeChar} returns a word (string) specifiying the datatype
of each column with a single character.
}
\author{
Akshay Gupta
}
