% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newdevice.R
\name{newdevice}
\alias{newdevice}
\title{Open a New Graphics Device}
\usage{
newdevice(width = 7, height = 7, ...)
}
\arguments{
\item{width}{the width (in inches) of the new device}

\item{height}{the height (in inches) of the new device}

\item{...}{additional arguments passed to the new device function}
}
\description{
Opens a new graphics device
}
\details{
Depending on the system, difference devices are better.
The windows device works fine (for now), only attempt to speed up
any other devices that we're going to be using.
We speed them up by getting rid of buffering.
}
\author{
Tom Elliott
}
