% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{readMetadata}
\alias{readMetadata}
\title{iNZight Metadata Reader}
\usage{
readMetadata(path)
}
\arguments{
\item{path}{A string. Specifies the location of the file to be read.}
}
\value{
A list.
}
\description{
Reads a file and checks for metadata if available.
}
\details{
\code{readMetadata} Returns the metadata if available in the file as a list.
}
\author{
Akshay Gupta
}
