% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{remove_rows}
\alias{remove_rows}
\title{Remove rows from data by row numbers}
\usage{
remove_rows(data, rows)
}
\arguments{
\item{data}{A dataframe or a survey design object to be filtered.}

\item{rows}{A numeric vector of row numbers to be sliced off.}
}
\value{
A filtered dataframe with the tidyverse code attached.
}
\description{
This function filters a dataframe or a survey design object by removing
specified rows based on the provided row numbers. The resulting filtered
dataframe is returned, along with the tidyverse code used to generate it.
}
\examples{
data <- remove_rows(iris, rows = c(1, 4, 5))
cat(code(data))
head(data)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin, Zhaoming Su
}
