% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{remSpacesAroundOperators}
\alias{remSpacesAroundOperators}
\title{Remove spaces around operators like '=','+','-', and back-slash.}
\usage{
remSpacesAroundOperators(string)
}
\arguments{
\item{string}{A string. A line in the metadata.}
}
\value{
A string.
}
\description{
\code{remSpacesAroundOperators} Returns a formatted string with 
spaces around operators listed removed.
}

