% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_levels.R
\name{rename.levels}
\alias{rename.levels}
\title{Renames the levels of a factor.}
\usage{
rename.levels(dafr, column, new.levels)
}
\arguments{
\item{dafr}{A data.frame of the data to change.}

\item{column}{The column name of the column to change.}

\item{new.levels}{A character variable of the length of 
the number of factors of the column to change. This 
vector contains the new levels.}
}
\value{
A data.frame where the levels of the specified 
columns are changed.
}
\description{
Renames the levels of a factor.
}
\author{
Christoph Knapp
}
