% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{rename_vars}
\alias{rename_vars}
\title{Rename column names}
\usage{
rename_vars(data, tobe_asis)
}
\arguments{
\item{data}{a dataframe with columns to rename}

\item{tobe_asis}{a named list of the old column names assigned
to the new column names
ie. list('new column names' = 'old column names')}
}
\value{
original dataframe containing new columns of the renamed columns
       with tidyverse code attached
}
\description{
Rename columns of a dataset with desired names
}
\examples{
renamed <- rename_vars(iris, list(
    sepal_length = "Sepal.Length",
    sepal_width = "Sepal.Width",
    petal_length = "Petal.Length",
    petal_width = "Petal.Width"
))
cat(code(renamed))
head(renamed)

}
\seealso{
\code{\link{code}}
}
\author{
Zhaoming Su
}
