% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{reshape_data}
\alias{reshape_data}
\title{Reshaping dataset from wide to long or from long to wide}
\usage{
reshape_data(.data, col1, col2, cols, key, value, check)
}
\arguments{
\item{.data}{dataset}

\item{col1}{column to spread out (for long to wide)}

\item{col2}{values to be put in the spread out column (for long to wide)}

\item{cols}{columns(s) to gather together (for wide to long)}

\item{key}{name for new column containing old column names (for wide to long)}

\item{value}{name for new column containing old column values
(for wide to long)}

\item{check}{check whether to use long to wide or wide to long}
}
\value{
reshaped dataset
}
\description{
Reshaping dataset from wide to long or from long to wide
}
\author{
Yiwen He
}
