% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_var.R
\name{separate_var}
\alias{separate_var}
\title{Separate columns}
\usage{
separate_var(data, var, by, left, right, into = c("cols", "rows"))
}
\arguments{
\item{data}{dataset}

\item{var}{name of variable to be separated}

\item{by}{a string as delimiter between values (separate by delimiter) or an
integer as number of characters to split by (separate by position)}

\item{left}{name for the separated left column (for \code{into = "cols"})}

\item{right}{name for the separated right column (for \code{into = "cols"})}

\item{into}{whether to split into new rows or columns}
}
\value{
Separated dataset
}
\description{
Separate columns
}
\author{
Stephen Su
}
