% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{sheets}
\alias{sheets}
\title{List available sheets within a file}
\usage{
sheets(x)
}
\arguments{
\item{x}{a dataframe, presumably returned by \code{smart_read}}
}
\value{
vector of sheet names, or NULL if the file was not an Excel workbook
}
\description{
Useful when reading an Excel file to quickly check what
other sheets are available.
}
\examples{
cas_file <- system.file("extdata/cas500.xls", package = "iNZightTools")
cas <- smart_read(cas_file)
sheets(cas)
}
\author{
Tom Elliott
}
