% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{smart_read}
\alias{smart_read}
\title{iNZight Smart Read}
\usage{
smart_read(
  file,
  ext = tools::file_ext(file),
  preview = FALSE,
  column_types = NULL,
  ...
)
}
\arguments{
\item{file}{the file path to read}

\item{ext}{file extension, namely "csv" or "txt"}

\item{preview}{logical, if \code{TRUE} only the first few rows of
the data will be returned}

\item{column_types}{vector of column types (see \code{?readr::read_csv})}

\item{...}{additional parameters passed to read_* functions}
}
\value{
a dataframe with attributes
}
\description{
A simple function that magically imports a file, irrespective of type.
}
\author{
Tom Elliott
}
