% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_data.R
\name{sort.data}
\alias{sort.data}
\title{Sort data by variables}
\usage{
\method{sort}{data}(.data, vars, asc = rep(TRUE, length(vars)))
}
\arguments{
\item{.data}{a dataframe to sort}

\item{vars}{a character vector of variable names to sort by}

\item{asc}{logical, same length as \code{vars}. 
If \code{TRUE} (default), sorted in ascending order, 
otherwise descending.}
}
\value{
data.frame with tidyverse code attached
}
\description{
Sorts a dataframe by one or more variables, and returns the result
along with tidyverse code used to generate it.
}
\examples{
sorted <- sort.data(iris, c("Sepal.Width", "Sepal.Length"), asc = c(TRUE, FALSE))
code(sorted)
head(sorted)

}
\seealso{
\code{\link{code}}
}
\author{
Owen Jin
}
