% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_variables.R
\name{stack.variables.perform}
\alias{stack.variables.perform}
\title{Stacks the selected columns onto the data.}
\usage{
\method{stack}{variables.perform}(columns, dafr)
}
\arguments{
\item{columns}{The columns to stack.}

\item{dafr}{a dataframe the stacking is performed on.}
}
\description{
Multiplies the data.set by adding rows to the data for 
every selected column. The selected columns are stacked
onto each other and added as an additional column
}
\author{
Christoph Knapp
}

